backend/requirements.txt
python-dotenv
openai-whisper
yt-dlp
twitch-dl